[OpTestPy] Add cumsum/cos/sin/ctc_align/cos_sim/crop/crop_tensor/correlation host utest test=develop,test=document_fix
