+++ lite/tests/unittest_py/op/test_cos_op.py
@@ -13,34 +13,53 @@
 # limitations under the License.
 
 import sys
-sys.path.append('../../common')
-sys.path.append('../../../')
+sys.path.append('../')
 
-import test_correlation_op_base
 from auto_scan_test import AutoScanTest, IgnoreReasons
 from program_config import TensorConfig, ProgramConfig, OpConfig, CxxConfig, TargetType, PrecisionType, DataLayoutType, Place
 import unittest
 
 import hypothesis
-from hypothesis import given, settings, seed, example, assume
+from hypothesis import given, settings, seed, example, assume, reproduce_failure
+import hypothesis.strategies as st
+import numpy as np
+from functools import partial
 
-class TestCorrelationOp(AutoScanTest):
+class TestCosOp(AutoScanTest):
+    def __init__(self, *args, **kwargs):
+        AutoScanTest.__init__(self, *args, **kwargs)
+        self.enable_testing_on_place(TargetType.Host, PrecisionType.FP32, DataLayoutType.NCHW, thread=[1,4])
+        self.enable_testing_on_place(TargetType.OpenCL, PrecisionType.FP16, DataLayoutType.ImageDefault)
+
     def is_program_valid(self, program_config: ProgramConfig , predictor_config: CxxConfig) -> bool:
         return True
 
     def sample_program_configs(self, draw):
-        return test_correlation_op_base.sample_program_configs(draw)
+        in_shape = draw(st.lists(st.integers(min_value=1, max_value=8), min_size=4, max_size=4))
+        cos_op = OpConfig(
+            type = "cos",
+            inputs = {"X" : ["input_data"]},
+            outputs = {"Out": ["output_data"]},
+            attrs = {})
+        program_config = ProgramConfig(
+            ops=[cos_op],
+            weights={},
+            inputs={
+                "input_data":
+                TensorConfig(shape=in_shape)
+            },
+            outputs=["output_data"])
+        return program_config
 
     def sample_predictor_configs(self):
         config = CxxConfig()
-        config.set_valid_places({Place(TargetType.Host, PrecisionType.FP32, DataLayoutType.NCHW)})
-        yield config, ["correlation"], (1e-5, 1e-5)
+        return self.get_predictor_configs(), ["cos"], (1e-5, 1e-5)
 
     def add_ignore_pass_case(self):
         pass
 
     def test(self, *args, **kwargs):
-        self.run_and_statis(quant=False, max_examples=25)
+        self.run_and_statis(quant=False, max_examples=300)
 
 if __name__ == "__main__":
     unittest.main(argv=[''])
